+++ pydbtools/__init__.py
@@ -1,7 +1,4 @@
-from .read_sql import read_sql
-from .get_athena_query_response import get_athena_query_response
-
-from .wrangler import (
+from ._wrangler import (  # noqa: F401
     create_temp_table,
     read_sql_query,
     read_sql_table,
@@ -16,10 +13,12 @@ from .wrangler import (
     stop_query_execution,
     wait_query,
     start_query_execution_and_wait,
-    create_temp_table,
     create_temp_database,
 )
 
-import poetry_version
+from ._sql_render import (  # noqa: F401
+    get_sql_from_file,
+    render_sql_template,
+)
 
-__version__ = poetry_version.extract(source_file=__file__)
+__version__ = "4.0.0"