+++ tests/test_wrangler.py
@@ -1,10 +1,10 @@
 import pytest
 
-from pydbtools.wrangler import init_athena_params
+from pydbtools._wrangler import init_athena_params
 
 
 def mock_get_user_id_and_table_dir(
-    boto3_session=None, force_ec2: bool = False, region_name: str = "eu-west-1"
+    boto3_session=None, force_ec2: bool = False, region_name: str = None
 ):
     if isinstance(boto3_session, dict) and "events" in boto3_session:
         boto3_session["events"].append("get_user_id_and_table_dir")
@@ -15,10 +15,12 @@ def mock_create_temp_database(
     temp_db_name: str = None,
     boto3_session=None,
     force_ec2: bool = False,
-    region_name: str = "eu-west-1",
+    region_name: str = None,
 ):
     if isinstance(boto3_session, dict) and "events" in boto3_session:
-        boto3_session["events"].append(f"_create_temp_database({temp_db_name})")
+        boto3_session["events"].append(
+            f"_create_temp_database({temp_db_name})"
+        )
     return boto3_session
 
 
@@ -135,16 +137,16 @@ def fun_with_sql_db_s3_ctas(
 def test_init_athena_params(
     fun_params, fun, expected_events, expect_warns, returned_db, monkeypatch
 ):
-    monkeypatch.setattr("pydbtools.wrangler.get_boto_session", get_empty_boto_log)
+    monkeypatch.setattr("pydbtools._wrangler.get_boto_session", get_empty_boto_log)
     monkeypatch.setattr(
-        "pydbtools.wrangler.get_user_id_and_table_dir", mock_get_user_id_and_table_dir
+        "pydbtools._wrangler.get_user_id_and_table_dir", mock_get_user_id_and_table_dir
     )
     monkeypatch.setattr(
-        "pydbtools.wrangler.get_database_name_from_userid",
+        "pydbtools._wrangler.get_database_name_from_userid",
         lambda user_id: "mojap_de_temp_pytest",
     )
     monkeypatch.setattr(
-        "pydbtools.wrangler._create_temp_database", mock_create_temp_database
+        "pydbtools._wrangler._create_temp_database", mock_create_temp_database
     )
 
     if expect_warns: